Tweak trial submission:
* add empty 3rd step (scoring template design)
* split last step into steps 5 (layout upload) and 6 (provenance)
