+++ app/forms/submissions/trial/step03_content_form.rb
@@ -1,22 +1,8 @@
 module Submissions
   module Trial
     class Step03ContentForm < PlantTrialForm
-      property :trait_scores, writeable: false
-      property :upload_id
-      property :trait_mapping, writeable: false
-
-      def self.permitted_properties
-        [
-          :upload_id,
-          :trait_mapping,
-          :trait_scores
-        ]
-      end
-
-      def upload
-        upload = Submission::Upload.trait_scores.find_by(id: upload_id)
-        SubmissionTraitScoresUploadDecorator.decorate(upload) if upload
-      end
+      # NOTE: needed because completely empty step breaks submission
+      property :_i_am_fake
     end
   end
 end
