+++ app/forms/submissions/trial/step04_content_form.rb
@@ -1,16 +1,21 @@
 module Submissions
   module Trial
     class Step04ContentForm < PlantTrialForm
-      property :layout_upload_id
-      property :visibility, default: 'published'
-      property :data_owned_by
-      property :data_provenance
-      property :comments
+      property :trait_scores, writeable: false
+      property :upload_id
+      property :trait_mapping, writeable: false
 
-      validates :visibility, inclusion: { in: %w(published private) }
+      def self.permitted_properties
+        [
+          :upload_id,
+          :trait_mapping,
+          :trait_scores
+        ]
+      end
 
-      def layout_upload
-        Submission::Upload.plant_trial_layout.find_by(id: layout_upload_id)
+      def upload
+        upload = Submission::Upload.trait_scores.find_by(id: upload_id)
+        SubmissionTraitScoresUploadDecorator.decorate(upload) if upload
       end
     end
   end
