+++ config/locales/submissions.en.yml
@@ -55,9 +55,15 @@ en:
             category: Select, or provide, the category of trait.
             units_of_measurements: Select, or write down, the unit of measurement when scoring this trait.
         step03:
+          heading: Scoring template design.
+          body: ...
+        step04:
           heading: Scoring.
           body: Upload tabular list of scores given to plants per each trait scored.
-        step04:
+        step05:
+          heading: Trial layout.
+          body: ...
+        step06:
           heading: Provenance.
           body: Identify the owner of the submitted data and describe its provenance.
       qtl:
