+++ spec/requests/submissions_spec.rb
@@ -126,7 +126,8 @@ RSpec.describe "Submission management" do
           end
 
           context "if last step" do
-            before { 3.times { submission.step_forward } }
+            let(:step_count) { submission.steps.count }
+            before { (step_count - 1).times { submission.step_forward } }
 
             context "and can finalize" do
               before { allow_any_instance_of(finalizer_klass).to receive(:call).and_return(true) }
